reorganize binsearch location
